web/agrolens-backend/requirements.txt
fastapi
uvicorn
pillow
numpy
tensorflow
transformers
torch
python-multipart
tf-keras
diskcache
